requirements.txt
yfinance
nltk
pandas
numpy
scikit-learn
tensorflow
pandas_datareader
newsapi-python
transformers
scipy
torch